+++ src/LoRaTNC.cpp
@@ -51,6 +51,7 @@ bool LoRaTNC::begin()
     // Set up KISS callbacks
     kissProtocol->onDataFrame(kissDataFrameWrapper);
     kissProtocol->onCommand(kissCommandWrapper);
+    kissProtocol->onLoRaCommand(kissLoRaCommandWrapper);
 
     // LoRa callbacks are assumed to be set up externally
     // since we need the LoRaRadio to call our methods
@@ -253,6 +254,204 @@ void LoRaTNC::handleKissCommand(uint8_t command, uint8_t parameter)
     //               KissProtocol::commandToString(command).c_str(), command, parameter);
 }
 
+void LoRaTNC::handleKissLoRaCommand(uint8_t command, uint8_t *data, uint16_t length)
+{
+    bool success = false;
+    uint8_t responseData[8];
+    uint16_t responseLength = 0;
+    
+    // Handle LoRa-specific KISS commands
+    switch (command)
+    {
+    case KISS_LORA_SET_TXPOWER:
+        if (length >= 1)
+        {
+            int8_t power = (int8_t)data[0];
+            int result = loraRadio->setTxPower(power);
+            success = (result == RADIOLIB_ERR_NONE);
+        }
+        break;
+        
+    case KISS_LORA_SET_BANDWIDTH:
+        if (length >= 1)
+        {
+            float bandwidth = KissProtocol::bandwidthIndexToValue(data[0]);
+            int result = loraRadio->setBandwidth(bandwidth);
+            success = (result == RADIOLIB_ERR_NONE);
+        }
+        break;
+        
+    case KISS_LORA_SET_SF:
+        if (length >= 1)
+        {
+            uint8_t sf = data[0];
+            if (sf >= 6 && sf <= 12)
+            {
+                int result = loraRadio->setSpreadingFactor(sf);
+                success = (result == RADIOLIB_ERR_NONE);
+            }
+        }
+        break;
+        
+    case KISS_LORA_SET_CR:
+        if (length >= 1)
+        {
+            uint8_t cr = data[0];
+            if (cr >= 5 && cr <= 8)
+            {
+                int result = loraRadio->setCodingRate(cr);
+                success = (result == RADIOLIB_ERR_NONE);
+            }
+        }
+        break;
+        
+    case KISS_LORA_SET_PREAMBLE:
+        if (length >= 2)
+        {
+            uint16_t preamble = (data[0] << 8) | data[1];
+            int result = loraRadio->setPreambleLength(preamble);
+            success = (result == RADIOLIB_ERR_NONE);
+        }
+        break;
+        
+    case KISS_LORA_SET_SYNCWORD:
+        if (length >= 1)
+        {
+            uint8_t syncWord = data[0];
+            int result = loraRadio->setSyncWord(syncWord);
+            success = (result == RADIOLIB_ERR_NONE);
+        }
+        break;
+        
+    case KISS_LORA_SET_CRC:
+        if (length >= 1)
+        {
+            bool enableCrc = (data[0] != 0);
+            int result = loraRadio->setCRC(enableCrc);
+            success = (result == RADIOLIB_ERR_NONE);
+        }
+        break;
+        
+    case KISS_LORA_SET_FREQ_LOW:
+        if (length >= 4)
+        {
+            uint32_t frequencyHz = (uint32_t(data[0]) << 24) | 
+                                  (uint32_t(data[1]) << 16) | 
+                                  (uint32_t(data[2]) << 8) | 
+                                  uint32_t(data[3]);
+            float frequencyMHz = frequencyHz / 1000000.0;
+            int result = loraRadio->setFrequency(frequencyMHz);
+            success = (result == RADIOLIB_ERR_NONE);
+        }
+        break;
+        
+    case KISS_LORA_SELECT_BAND:
+        if (length >= 1)
+        {
+            uint8_t bandIndex = data[0];
+            // Get available bands from the frequency band manager
+            FrequencyBandManager* bandManager = loraRadio->getBandManager();
+            if (bandManager)
+            {
+                auto bands = bandManager->getAvailableBands();
+                if (bandIndex < bands.size())
+                {
+                    success = bandManager->selectBand(bands[bandIndex].identifier);
+                    if (success)
+                    {
+                        // Set the radio frequency to the band's default
+                        loraRadio->setFrequency(bands[bandIndex].defaultFrequency);
+                    }
+                }
+            }
+        }
+        break;
+        
+    case KISS_LORA_GET_CONFIG:
+        {
+            // Send current configuration as response
+            lora_config_t config = loraRadio->getConfig();
+            
+            // Send frequency (4 bytes)
+            uint32_t freqHz = (uint32_t)(config.frequency * 1000000);
+            responseData[0] = (freqHz >> 24) & 0xFF;
+            responseData[1] = (freqHz >> 16) & 0xFF;
+            responseData[2] = (freqHz >> 8) & 0xFF;
+            responseData[3] = freqHz & 0xFF;
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_FREQ_LOW, responseData, 4);
+            
+            // Send TX power
+            responseData[0] = (uint8_t)config.txPower;
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_TXPOWER, responseData, 1);
+            
+            // Send bandwidth
+            responseData[0] = KissProtocol::bandwidthValueToIndex(config.bandwidth);
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_BANDWIDTH, responseData, 1);
+            
+            // Send spreading factor
+            responseData[0] = config.spreadingFactor;
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_SF, responseData, 1);
+            
+            // Send coding rate
+            responseData[0] = config.codingRate;
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_CR, responseData, 1);
+            
+            // Send preamble length
+            responseData[0] = (config.preambleLength >> 8) & 0xFF;
+            responseData[1] = config.preambleLength & 0xFF;
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_PREAMBLE, responseData, 2);
+            
+            // Send sync word
+            responseData[0] = config.syncWord;
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_SYNCWORD, responseData, 1);
+            
+            // Send CRC setting
+            responseData[0] = config.crcEnabled ? 1 : 0;
+            kissProtocol->sendLoRaResponse(KISS_LORA_SET_CRC, responseData, 1);
+            
+            success = true;
+        }
+        break;
+        
+    case KISS_LORA_SAVE_CONFIG:
+        {
+            // Save configuration to NVS (if implemented)
+            FrequencyBandManager* bandManager = loraRadio->getBandManager();
+            if (bandManager)
+            {
+                success = bandManager->saveConfiguration();
+            }
+        }
+        break;
+        
+    case KISS_LORA_RESET_CONFIG:
+        {
+            // Reset to default configuration
+            loraRadio->setFrequency(915.0);         // Default frequency
+            loraRadio->setTxPower(LORA_DEFAULT_TX_POWER);
+            loraRadio->setBandwidth(LORA_BANDWIDTH_DEFAULT);
+            loraRadio->setSpreadingFactor(LORA_SPREADING_FACTOR_DEFAULT);
+            loraRadio->setCodingRate(LORA_CODINGRATE_DEFAULT);
+            loraRadio->setPreambleLength(LORA_PREAMBLE_LENGTH_DEFAULT);
+            loraRadio->setSyncWord(LORA_SYNC_WORD_DEFAULT);
+            loraRadio->setCRC(true);
+            success = true;
+        }
+        break;
+        
+    default:
+        // Unknown LoRa command - silently ignore
+        break;
+    }
+    
+    // For commands that don't send their own responses, send a simple ACK/NAK
+    if (command != KISS_LORA_GET_CONFIG)
+    {
+        responseData[0] = success ? 0x01 : 0x00;  // 1 = success, 0 = failure
+        kissProtocol->sendLoRaResponse(command, responseData, 1);
+    }
+}
+
 void LoRaTNC::handleLoRaReceive(uint8_t *payload, uint16_t size, int16_t rssi, float snr)
 {
     stats.packetsReceived++;
@@ -400,5 +599,13 @@ void LoRaTNC::kissCommandWrapper(uint8_t command, uint8_t parameter)
     if (instance)
     {
         instance->handleKissCommand(command, parameter);
+    }
+}
+
+void LoRaTNC::kissLoRaCommandWrapper(uint8_t command, uint8_t *data, uint16_t length)
+{
+    if (instance)
+    {
+        instance->handleKissLoRaCommand(command, data, length);
     }
 }